Package window.go
