+++ window/window_test.go
@@ -1,4 +1,4 @@
-package main
+package window
 
 import (
 	"reflect"
@@ -25,28 +25,5 @@ func TestMakeLineNum(t *testing.T) {
 	exp = []rune{'1', '2', '3', '4', ' '}
 	if !reflect.DeepEqual(got, exp) {
 		t.Errorf("failed : got : %v, exp : %v", got, exp)
-	}
-}
-
-func TestGetDigits(t *testing.T) {
-	got := getDigit(1)
-	if got != 1 {
-		t.Errorf("failed")
-	}
-	got = getDigit(12)
-	if got != 2 {
-		t.Errorf("failed")
-	}
-	got = getDigit(123)
-	if got != 3 {
-		t.Errorf("failed")
-	}
-	got = getDigit(1234)
-	if got != 4 {
-		t.Errorf("failed")
-	}
-	got = getDigit(12345)
-	if got != 5 {
-		t.Errorf("failed")
 	}
 }
